Move ChannelBanner to shared components and extract reusable observable
