+++ app/components/channel_banner/channel_banner.tsx
