+++ app/components/channel_banner/index.ts
@@ -6,7 +6,7 @@ import {of as of$} from 'rxjs';
 import {switchMap} from 'rxjs/operators';
 
 import {observeChannel} from '@queries/servers/channel';
-import {ChannelBanner} from '@screens/channel/header/channel_banner/channel_banner';
+import {ChannelBanner} from '@components/channel_banner/channel_banner';
 
 import type {WithDatabaseArgs} from '@typings/database/database';
