+++ app/screens/thread/index.tsx
@@ -2,16 +2,15 @@
 // See LICENSE.txt for license information.
 
 import {withDatabase, withObservables} from '@nozbe/watermelondb/react';
-import {combineLatestWith, distinctUntilChanged, switchMap, of as of$} from 'rxjs';
+import {distinctUntilChanged, switchMap, of as of$} from 'rxjs';
 
 import {observeCallStateInChannel} from '@calls/observers';
 import {withServerUrl} from '@context/server';
 import {observeChannel} from '@queries/servers/channel';
 import {observePost} from '@queries/servers/post';
 import {observeScheduledPostCountForThread} from '@queries/servers/scheduled_post';
-import {observeLicense} from '@queries/servers/system';
 import {observeIsCRTEnabled} from '@queries/servers/thread';
-import {shouldShowChannelBanner} from '@screens/channel/channel_feature_checks';
+import {observeChannelBannerIncluded} from '@screens/channel/channel_feature_checks';
 import EphemeralStore from '@store/ephemeral_store';
 
 import Thread from './thread';
@@ -40,18 +39,7 @@ const enhanced = withObservables(['rootId'], ({database, serverUrl, rootId}: Enh
         switchMap((c) => of$(c?.type)),
     );
 
-    const bannerInfo = channel.pipe(
-        switchMap((c) => of$(c?.bannerInfo)),
-    );
-
-    const license = observeLicense(database);
-
-    const includeChannelBanner = channelType.pipe(
-        combineLatestWith(license, bannerInfo),
-        switchMap(([channelTypeValue, licenseValue, bannerInfoValue]) =>
-            of$(shouldShowChannelBanner(channelTypeValue, licenseValue, bannerInfoValue)),
-        ),
-    );
+    const includeChannelBanner = observeChannelBannerIncluded(database, channelType, channelId);
 
     const scheduledPostCount = observeScheduledPostCountForThread(database, rootId);
